+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.2.0] - 2025-07-18
+
+### Changed
+
+- Improve `HighlightController` to limit highlighting on large files and added error handling during keyword parsing
+- Refactor `CommentService` to support new languages with appropriate comment templates
+- Enhance `escapeRegExp` function for better regex handling and moved it to a dedicated helper file
+
 ## [1.1.1] - 2025-03-09
 
 ### Fixed
@@ -36,7 +44,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release of codeMark+ extension
 
-[Unreleased]: https://github.com/ManuelGil/vscode-code-mark-plus/compare/v1.1.1...HEAD
+[Unreleased]: https://github.com/ManuelGil/vscode-code-mark-plus/compare/v1.2.0...HEAD
+[1.2.0]: https://github.com/ManuelGil/vscode-code-mark-plus/compare/v1.1.1...v1.2.0
 [1.1.1]: https://github.com/ManuelGil/vscode-code-mark-plus/compare/v1.1.0...v1.1.1
 [1.1.0]: https://github.com/ManuelGil/vscode-code-mark-plus/compare/v1.0.1...v1.1.0
 [1.0.1]: https://github.com/ManuelGil/vscode-code-mark-plus/compare/v1.0.0...v1.0.1